requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic[email]
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
python-dotenv~=1.0.1
mysql-connector-python~=9.2.0
email-validator
bcrypt==4.0.1
passlib==1.7.4
pydantic-settings~=2.7.1